long long

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -14,7 +14,7 @@ int main(void)
 
 	int two = 2;
 
-	unsigned long int sum = one + two;
+	unsigned long long int sum = one + two;
 
 	printf("%d, ", one);
 	printf("%d, ", two);
@@ -26,10 +26,10 @@ int main(void)
 		two = sum;
 		if (i == 47)
 		{
-			printf("%lu", sum);
+			printf("%llu", sum);
 			break;
 		}
-		printf("%lu, ", sum);
+		printf("%llu, ", sum);
 
 	}
 	printf("\n");
